# Possible optimization for blit functions