@itzpr3d4t0r 2023-06-10
Changed test program again to include one more case, here is the new one and new results:
**WITH THE CHANGES**
```
OUTSIDE
fblits) 0.036678159999974015
blits doreturn) 0.1063232300000891
blits) 0.08050288000013098
blit) 0.00020791000006283866

INSIDE
fblits) 0.605369150000115
blits doreturn) 0.7010963499998979
blits) 0.6715423200004806
blit) 0.0008256900002379553
```
**MAIN**
```
OUTSIDE
fblits) 0.13856487999983075
blits doreturn) 0.208694860000287
blits) 0.18958259999963048
blit) 0.0003108700000666431

INSIDE
fblits) 0.5962242999994487
blits doreturn) 0.703472459999648
blits) 0.6794426200000089
blit) 0.000789950000216777
```

```Python
import pygame
from timeit import repeat
from statistics import mean


pygame.init()
WIDTH, HEIGHT = 400, 400
win = pygame.display.set_mode((WIDTH, HEIGHT))

surf = pygame.Surface((WIDTH, HEIGHT))
surf.fill((255, 0, 0))

lst_inside = [
    (surf, (350, 350)) for _ in range(1000)
]

lst_outside = [
    (surf, (1000, 1000)) for _ in range(1000)
]

G = globals()

def test(name: str, number=1000, rep=10) -> float:
    return mean(repeat(name, globals=G, number=number, repeat=rep))

print("OUTSIDE")
print("fblits)", test("win.fblits(lst_outside)"))
print("blits doreturn)", test("win.blits(lst_outside)"))
print("blits)", test("win.blits(lst_outside, False)"))
print("blit)", test("win.blit(surf, (1000, 1000))"))

print("\nINSIDE")
print("fblits)", test("win.fblits(lst_inside)"))
print("blits doreturn)", test("win.blits(lst_inside)"))
print("blits)", test("win.blits(lst_inside, False)"))
print("blit)", test("win.blit(surf, (350, 350))"))
```