@Starbuck5 2023-06-18
That is quite an interesting result, cool work.

SDL_BlitSurface and pygame_Blit both do bounds checking, so it's not like the blit is happening, it's just the set up and tear down of the blit operation that is taking time.

In SDL, there is SDL_UpperBlit and SDL_LowerBlit. UpperBlit is the same as BlitSurface, and does bounds checking. LowerBlit does no bounds checking.

So maybe a good approach to implement this would be to make pgSurface_Blit do the early exit bounds checking itself, then assuming it has a blit to do, call a new function pgSurface_BlitLower. This function could call `SDL_LowerBlit` or a new function `pygame_BlitLower` (`pygame_Blit` without bounds checking) so that it doesn't replicate the bounds check.