@MyreMylar 2023-10-12
I did a quick test swapping in `SDL_LowerBlit()` for `SDL_BlitSurface()` in this PR branch, but it segfaults so clearly `SDL_BlitSurface()` is still doing something useful we are not capturing at the higher level yet.